Исправил вроде. Но у меня работало всё и до этого

+++ lab4/ex3.py
@@ -106,7 +106,7 @@ class Ball():
         screen - экран
 
         '''
-        circle(screen, self.color, (self.x, self.y), self.r)
+        circle(screen, self.color, (int(self.x), int(self.y)), int(self.r))
 
     def new_score(self, x_m, y_m, ball, balls):
         '''
@@ -203,7 +203,7 @@ class Square():
 
         '''
         a = self.a
-        rect(screen, self.color, (self.x - a / 2, self.y - a / 2, a, a))
+        rect(screen, self.color, (int(self.x - a / 2), int(self.y - a / 2), int(a), int(a)))
 
     def new_score(self, x_m, y_m, square, squares):
         '''
